# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NCS Warehouse Management API",
        "description": "API untuk manajemen gudang PT. NCS Bandung dengan algoritma Simulated Annealing",
        "contact": {
            "name": "NCS Development Team",
            "email": "dev@ncs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/barang/{id}": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Detail barang",
                "description": "Mengambil detail barang berdasarkan ID",
                "operationId": "showBarang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail barang berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail barang berhasil diambil"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Barang"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Barang tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Barang"
                ],
                "summary": "Update barang",
                "description": "Mengupdate data barang",
                "operationId": "updateBarang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_barang": {
                                        "type": "string",
                                        "example": "Laptop Acer Aspire 5 Updated"
                                    },
                                    "kategori_barang_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "panjang": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 36
                                    },
                                    "lebar": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 26
                                    },
                                    "tinggi": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    },
                                    "berat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2
                                    },
                                    "mudah_pecah": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "prioritas": {
                                        "type": "string",
                                        "enum": [
                                            "rendah",
                                            "sedang",
                                            "tinggi"
                                        ],
                                        "example": "tinggi"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Laptop untuk kebutuhan kantor yang sudah diupdate"
                                    },
                                    "aktif": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barang berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Barang berhasil diupdate"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Barang"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "Barang tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Barang"
                ],
                "summary": "Hapus barang",
                "description": "Menghapus barang (soft delete jika masih ada penempatan)",
                "operationId": "deleteBarang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barang berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Barang berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Barang tidak ditemukan"
                    },
                    "409": {
                        "description": "Barang masih memiliki penempatan aktif"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/barang/{id}/toggle-status": {
            "patch": {
                "tags": [
                    "Barang"
                ],
                "summary": "Toggle status aktif barang",
                "description": "Mengubah status aktif/nonaktif barang",
                "operationId": "toggleBarangStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status barang berhasil diubah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Status barang berhasil diubah"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Barang"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Barang tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}